+++ wandb/sdk/interface/interface.py
@@ -519,6 +519,7 @@ class InterfaceBase:
         run: "Run",
         artifact: "Artifact",
         aliases: Iterable[str],
+        tags: Optional[Iterable[str]] = None,
         history_step: Optional[int] = None,
         is_user_created: bool = False,
         use_after_commit: bool = False,
@@ -532,8 +533,9 @@ class InterfaceBase:
         proto_artifact.user_created = is_user_created
         proto_artifact.use_after_commit = use_after_commit
         proto_artifact.finalize = finalize
-        for alias in aliases:
-            proto_artifact.aliases.append(alias)
+
+        proto_artifact.aliases.extend(aliases or [])
+        proto_artifact.tags.extend(tags or [])
 
         log_artifact = pb.LogArtifactRequest()
         log_artifact.artifact.CopyFrom(proto_artifact)
@@ -577,6 +579,7 @@ class InterfaceBase:
         run: "Run",
         artifact: "Artifact",
         aliases: Iterable[str],
+        tags: Optional[Iterable[str]] = None,
         is_user_created: bool = False,
         use_after_commit: bool = False,
         finalize: bool = True,
@@ -589,8 +592,8 @@ class InterfaceBase:
         proto_artifact.user_created = is_user_created
         proto_artifact.use_after_commit = use_after_commit
         proto_artifact.finalize = finalize
-        for alias in aliases:
-            proto_artifact.aliases.append(alias)
+        proto_artifact.aliases.extend(aliases or [])
+        proto_artifact.tags.extend(tags or [])
         self._publish_artifact(proto_artifact)
 
     @abstractmethod